+++ packages/core/src/ws/models.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 /// 消息类型的信令映射枚举
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[repr(u8)]
+#[repr(i32)]
 pub enum CallMsgType {
     Heartbeat = 2,
     VideoCallRequest = 5,
@@ -14,10 +14,10 @@ pub enum CallMsgType {
     Dropped = 23,
 }
 
-impl TryFrom<u8> for CallMsgType {
+impl TryFrom<i32> for CallMsgType {
     type Error = String;
 
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
+    fn try_from(value: i32) -> Result<Self, Self::Error> {
         match value {
             2 => Ok(Self::Heartbeat),
             5 => Ok(Self::VideoCallRequest),
@@ -35,31 +35,31 @@ impl TryFrom<u8> for CallMsgType {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct WsBaseResp {
     #[serde(rename = "type")]
-    pub msg_type: u8,
+    pub msg_type: i32,
     pub data: serde_json::Value,
 }
 
 /// 请求通话数据
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct CallRequestData {
-    pub target_uid: u64,
-    pub room_id: u64,
+    pub target_uid: i64,
+    pub room_id: i64,
     pub is_video: bool,
 }
 
 /// 响应通话数据
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct CallResponseData {
-    pub caller_uid: u64,
-    pub room_id: u64,
-    pub accepted: u8, // 1 for yes, 0 for no
+    pub caller_uid: i64,
+    pub room_id: i64,
+    pub accepted: i32, // 1 for yes, 0 for no
 }
 
 /// 接收到的来电数据
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct IncomingCallData {
-    pub caller_uid: u64,
-    pub room_id: u64,
+    pub caller_uid: i64,
+    pub room_id: i64,
     pub is_video: bool,
 }
 
@@ -73,7 +73,7 @@ pub struct CallAcceptedData {
 /// 用于取消或丢弃的通用 Room ID 结构体
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct CallRoomIdData {
-    pub room_id: u64,
+    pub room_id: i64,
 }
 
 /// WebSocket 客户端连接状态
